requirements.txt
asgiref==3.8.1
attrs==24.2.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.0.9
django-cleanup==9.0.0
django-cors-headers==4.5.0
django-filter==24.3
django-unfold==0.40.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
flatbuffers==24.3.25
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
numpy==2.1.2
openmeteo_requests==1.3.0
openmeteo_sdk==1.16.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.3
whitenoise==6.7.0
